src/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
gunicorn==23.0.0
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
packaging==24.1
psycopg2-binary==2.9.9
pyserial==3.5
python-dotenv==1.0.1
Werkzeug==3.0.3
zipp==3.20.0
flask
gunicorn
subprocess.run
psutil
python-nmap
requests
beautifulsoup4
flask 
python-dotenv
pymongo[srv]
pymongo
pytest
paramiko
flask-cors
gspread 
oauth2client
redis
flask-socketio
# Flask e componenti principali
Flask==3.0.3
flask-cors==4.0.0
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
blinker==1.8.2
click==8.1.7

# Server WSGI
gunicorn==23.0.0

# Database
mysql-connector-python==9.0.0
psycopg2-binary==2.9.9
pymongo[srv]==4.5.0

# Networking e sistema
python-nmap==0.7.1
psutil==5.9.5
paramiko==3.3.1

# Google Sheets e API Google
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
gspread==5.11.3
oauth2client==4.1.3

# Caching
redis==4.6.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
packaging==24.1
zipp==3.20.0
importlib_metadata==8.2.0
pyserial==3.5

# Realtime
flask-socketio==5.3.6

# Sviluppo e test
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0
